Do not hide any header ancestor in side-by-side view
Journal's article page and also the pattern-library page `pages-article--research` have a more complex markup than the original test. Due to this nesting of $header, one of its ancestors is hidden, hiding in turn its children: the header itself and the close bar of the side-by-side view.

Very open to feedback about how to best check the condition.