More use of font size and line height mixin
Now used for line heights everywhere outside mixins (and sometimes there too). In a handful of cases this will result in a redundant font-size declaration in the generated css, but it's highly compressable, so shouldn't have any noticeable effect for the user. In fact the ungzipped file size of the generated  all.css is 3kb smaller from this commit compared to the current head of develop 6bc0a5.