Line height mixin (#774)
* Add mixin to set font size and line height in the same step

* Refactor to font size and line height mixin where both font size and line height are expressed in the same block

* More use of font size and line height mixin

Now used for line heights everywhere outside mixins (and sometimes there too). In a handful of cases this will result in a redundant font-size declaration in the generated css, but it's highly compressable, so shouldn't have any noticeable effect for the user. In fact the ungzipped file size of the generated  all.css is 3kb smaller from this commit compared to the current head of develop 6bc0a5.

* Use the correct sass variable!

* Use $font-size-base-in-px in place of bare number 16 in all font sizing

* Bug fix

* Remove obsolete sass variable