Error: failed to parse stage name "node:": invalid reference format
Reason: Dockerfile ARG handling is complicated in a multi-stage file. For the Arg to be used for the FROM Commands, they need to be defined at the top level, for them to be used in a stage, they need defining after the FROM command. To use in both, requires both. See https://docs.docker.com/engine/reference/builder/#understand-how-arg-and-from-interact

Solution: Define build-arg args at the top of the file and define args where they are used.